Improve channel funding scheme - instead of deriving from current msec timestamp a random private key is used as funding key - funding public key also acts as seed for channel key path which is used to generate channel secrets
With this new scheme we only need a seed to recover a channel because once remote peer publishes a commit on our request we can extract our funding pubKey from it and use that key to re-derive all channel secrets needed to get our balance back. We still need to know which peer we had a channel with for this to work.